Create index.html
The purpose of this project was to create growing circles on my website using jQuery's animate. The index page simply creates a div styled by CSS into a circle for which jQuery targets the DOM element and performs an animate upon to grow and then shrink the circle to the desired size. 

I also refer to Google's minified copy of jQuery along with my locally stored copy as a fallback option, in case external sources fail.